Refactored Gettingstarted reducers
