+++ src/constants/gettingStarted.ts
@@ -1,0 +1,3 @@
+export default {
+  UPDATE: 'GETTINGSTARTED_UPDATE',
+}
