+++ src/main.tsx
@@ -11,7 +11,8 @@ import {Provider} from 'react-redux'
 import * as thunk from 'redux-thunk'
 import * as cookiestore from 'cookiestore'
 import drumstick from 'drumstick'
-import { reduceGettingStartedState, fetchGettingStartedState } from './reducers/GettingStartedState'
+import { reduceGettingStartedState } from './reducers/gettingStarted'
+import { fetchGettingStartedState } from './actions/gettingStarted'
 import { reducePopup } from './reducers/popup'
 import { reduceProgress } from './reducers/progressIndicator'
