+++ src/views/ProjectRootView/SideNav.tsx
@@ -13,7 +13,7 @@ import Tether from '../../components/Tether/Tether'
 import AddModelMutation from '../../mutations/AddModelMutation'
 import {sideNavSyncer} from '../../utils/sideNavSyncer'
 import {onFailureShowNotification} from '../../utils/relay'
-import {nextStep, skip} from '../../reducers/GettingStartedState'
+import {nextStep, skip} from '../../actions/gettingStarted'
 import {Project, Viewer, Model} from '../../types/types'
 import {ShowNotificationCallback} from '../../types/utils'
 import {classnames} from '../../utils/classnames'
